Enum winconsole::console::CodePage

#[repr(u16)]

pub enum CodePage {
    None,
    IBM037,
    IBM437,
    IBM500,
    ASMO_708,
    ASMO_709,
    ASMO_710,
    DOS_720,
    ibm737,
    ibm775,
    ibm850,
    ibm852,
    IBM855,
    ibm857,
    IBM00858,
    IBM860,
    ibm861,
    DOS_862,
    IBM863,
    IBM864,
    IBM865,
    cp866,
    ibm869,
    IBM870,
    windows_874,
    cp875,
    shift_jis,
    gb2312,
    ks_c_5601_1987,
    big5,
    IBM1026,
    IBM01047,
    IBM01140,
    IBM01141,
    IBM01142,
    IBM01143,
    IBM01144,
    IBM01145,
    IBM01146,
    IBM01147,
    IBM01148,
    IBM01149,
    utf_16,
    unicodeFFFE,
    windows_1250,
    windows_1251,
    windows_1252,
    windows_1253,
    windows_1254,
    windows_1255,
    windows_1256,
    windows_1257,
    windows_1258,
    Johab,
    macintosh,
    x_mac_japanese,
    x_mac_chinesetrad,
    x_mac_korean,
    x_mac_arabic,
    x_mac_hebrew,
    x_mac_greek,
    x_mac_cyrillic,
    x_mac_chinesesimp,
    x_mac_romanian,
    x_mac_ukrainian,
    x_mac_thai,
    x_mac_ce,
    x_mac_icelandic,
    x_mac_turkish,
    x_mac_croatian,
    utf_32,
    utf_32BE,
    x_Chinese_CNS,
    x_cp20001,
    x_Chinese_Eten,
    x_cp20003,
    x_cp20004,
    x_cp20005,
    x_IA5,
    x_IA5_German,
    x_IA5_Swedish,
    x_IA5_Norwegian,
    us_ascii,
    x_cp20261,
    x_cp20269,
    IBM273,
    IBM277,
    IBM278,
    IBM280,
    IBM284,
    IBM285,
    IBM290,
    IBM297,
    IBM420,
    IBM423,
    IBM424,
    x_EBCDIC_KoreanExtended,
    IBM_Thai,
    koi8_r,
    IBM871,
    IBM880,
    IBM905,
    IBM00924,
    EUC_JP,
    x_cp20936,
    x_cp20949,
    cp1025,
    koi8_u,
    iso_8859_1,
    iso_8859_2,
    iso_8859_3,
    iso_8859_4,
    iso_8859_5,
    iso_8859_6,
    iso_8859_7,
    iso_8859_8,
    iso_8859_9,
    iso_8859_13,
    iso_8859_15,
    x_Europa,
    iso_8859_8_i,
    iso_2022_jp_1,
    csISO2022JP,
    iso_2022_jp_2,
    iso_2022_kr,
    x_cp50227,
    iso_2022_ch,
    x_EBCDIC_JapaneseExtended,
    x_EBCDIC_USCanadaJapanese,
    x_EBCDIC_KoreanExtendedAndKorean,
    x_EBCDIC_ChineseExtendedAndChinese,
    x_EBCDIC_ChineseExtended,
    x_EBCDIC_USCanadaChinese,
    x_EBCDIC_JapaneseExtendedAndJapanese,
    euc_jp,
    EUC_CN,
    euc_kr,
    EUC_TCN,
    hz_gb_2312,
    GB18030,
    x_iscii_de,
    x_iscii_be,
    x_iscii_ta,
    x_iscii_te,
    x_iscii_as,
    x_iscii_or,
    x_iscii_ka,
    x_iscii_ma,
    x_iscii_gu,
    x_iscii_pa,
    utf_7,
    utf_8,
    Invalid,
}

Represents a console code page.

Variants
None

Default value.

IBM037

IBM EBCDIC US-Canada.

IBM437

OEM United States.

IBM500

IBM EBCDIC International.

ASMO_708

Arabic (ASMO 708).

ASMO_709

Arabic (ASMO-449+, BCON V4).

ASMO_710

Arabic - Transparent Arabic.

DOS_720

Arabic (Transparent ASMO); Arabic (DOS).

ibm737

OEM Greek (formerly 437G); Greek (DOS).

ibm775

OEM Baltic; Baltic (DOS).

ibm850

OEM Multilingual Latin 1; Western European (DOS).

ibm852

OEM Latin 2; Central European (DOS).

IBM855

OEM Cyrillic (primarily Russian).

ibm857

OEM Turkish; Turkish (DOS).

IBM00858

OEM Multilingual Latin 1 + Euro symbol.

IBM860

OEM Portuguese; Portuguese (DOS).

ibm861

OEM Icelandic; Icelandic (DOS).

DOS_862

OEM Hebrew; Hebrew (DOS).

IBM863

OEM French Canadian; French Canadian (DOS).

IBM864

OEM Arabic; Arabic (864).

IBM865

OEM Nordic; Nordic (DOS).

cp866

OEM Russian; Cyrillic (DOS).

ibm869

OEM Modern Greek; Greek, Modern (DOS).

IBM870

IBM EBCDIC Multilingual/ROECE (Latin 2); IBM EBCDIC Multilingual Latin 2.

windows_874

ANSI/OEM Thai (ISO 8859-11); Thai (Windows).

cp875

IBM EBCDIC Greek Modern.

shift_jis

ANSI/OEM Japanese; Japanese (Shift-JIS).

gb2312

ANSI/OEM Simplified Chinese (PRC, Singapore); Chinese Simplified (GB2312).

ks_c_5601_1987

ANSI/OEM Korean (Unified Hangul Code).

big5

ANSI/OEM Traditional Chinese (Taiwan; Hong Kong SAR, PRC); Chinese Traditional (Big5).

IBM1026

IBM EBCDIC Turkish (Latin 5).

IBM01047

IBM EBCDIC Latin 1/Open System.

IBM01140

IBM EBCDIC US-Canada (037 + Euro symbol); IBM EBCDIC (US-Canada-Euro).

IBM01141

IBM EBCDIC Germany (20273 + Euro symbol); IBM EBCDIC (Germany-Euro).

IBM01142

IBM EBCDIC Denmark-Norway (20277 + Euro symbol); IBM EBCDIC (Denmark-Norway-Euro).

IBM01143

IBM EBCDIC Finland-Sweden (20278 + Euro symbol); IBM EBCDIC (Finland-Sweden-Euro).

IBM01144

IBM EBCDIC Italy (20280 + Euro symbol); IBM EBCDIC (Italy-Euro).

IBM01145

IBM EBCDIC Latin America-Spain (20284 + Euro symbol); IBM EBCDIC (Spain-Euro).

IBM01146

IBM EBCDIC United Kingdom (20285 + Euro symbol); IBM EBCDIC (UK-Euro).

IBM01147

IBM EBCDIC France (20297 + Euro symbol); IBM EBCDIC (France-Euro).

IBM01148

IBM EBCDIC International (500 + Euro symbol); IBM EBCDIC (International-Euro).

IBM01149

IBM EBCDIC Icelandic (20871 + Euro symbol); IBM EBCDIC (Icelandic-Euro).

utf_16

Unicode UTF-16, little endian byte order (BMP of ISO 10646); available only to managed applications.

unicodeFFFE

Unicode UTF-16, big endian byte order; available only to managed applications.

windows_1250

ANSI Central European; Central European (Windows).

windows_1251

ANSI Cyrillic; Cyrillic (Windows).

windows_1252

ANSI Latin 1; Western European (Windows).

windows_1253

ANSI Greek; Greek (Windows).

windows_1254

ANSI Turkish; Turkish (Windows).

windows_1255

ANSI Hebrew; Hebrew (Windows).

windows_1256

ANSI Arabic; Arabic (Windows).

windows_1257

ANSI Baltic; Baltic (Windows).

windows_1258

ANSI/OEM Vietnamese; Vietnamese (Windows).

Johab

Korean (Johab).

macintosh

MAC Roman; Western European (Mac).

x_mac_japanese

Japanese (Mac).

x_mac_chinesetrad

MAC Traditional Chinese (Big5); Chinese Traditional (Mac).

x_mac_korean

Korean (Mac).

x_mac_arabic

Arabic (Mac).

x_mac_hebrew

Hebrew (Mac).

x_mac_greek

Greek (Mac).

x_mac_cyrillic

Cyrillic (Mac).

x_mac_chinesesimp

MAC Simplified Chinese (GB 2312); Chinese Simplified (Mac).

x_mac_romanian

Romanian (Mac).

x_mac_ukrainian

Ukrainian (Mac).

x_mac_thai

Thai (Mac).

x_mac_ce

MAC Latin 2; Central European (Mac).

x_mac_icelandic

Icelandic (Mac).

x_mac_turkish

Turkish (Mac).

x_mac_croatian

Croatian (Mac).

utf_32

Unicode UTF-32, little endian byte order; available only to managed applications.

utf_32BE

Unicode UTF-32, big endian byte order; available only to managed applications.

x_Chinese_CNS

CNS Taiwan; Chinese Traditional (CNS).

x_cp20001

TCA Taiwan.

x_Chinese_Eten

Eten Taiwan; Chinese Traditional (Eten).

x_cp20003

IBM5550 Taiwan.

x_cp20004

TeleText Taiwan.

x_cp20005

Wang Taiwan.

x_IA5

IA5 (IRV International Alphabet No. 5, 7-bit); Western European (IA5).

x_IA5_German

IA5 German (7-bit).

x_IA5_Swedish

IA5 Swedish (7-bit).

x_IA5_Norwegian

IA5 Norwegian (7-bit).

us_ascii

US-ASCII (7-bit).

x_cp20261

T.61.

x_cp20269

ISO 6937 Non-Spacing Accent.

IBM273

IBM EBCDIC Germany.

IBM277

IBM EBCDIC Denmark-Norway.

IBM278

IBM EBCDIC Finland-Sweden.

IBM280

IBM EBCDIC Italy.

IBM284

IBM EBCDIC Latin America-Spain.

IBM285

IBM EBCDIC United Kingdom.

IBM290

IBM EBCDIC Japanese Katakana Extended.

IBM297

IBM EBCDIC France.

IBM420

IBM EBCDIC Arabic.

IBM423

IBM EBCDIC Greek.

IBM424

IBM EBCDIC Hebrew.

x_EBCDIC_KoreanExtended

IBM EBCDIC Korean Extended.

IBM_Thai

IBM EBCDIC Thai.

koi8_r

Russian (KOI8-R); Cyrillic (KOI8-R).

IBM871

IBM EBCDIC Icelandic.

IBM880

IBM EBCDIC Cyrillic Russian.

IBM905

IBM EBCDIC Turkish.

IBM00924

IBM EBCDIC Latin 1/Open System (1047 + Euro symbol).

EUC_JP

Japanese (JIS 0208-1990 and 0212-1990).

x_cp20936

Simplified Chinese (GB2312); Chinese Simplified (GB2312-80).

x_cp20949

Korean Wansung.

cp1025

IBM EBCDIC Cyrillic Serbian-Bulgarian.

koi8_u

Ukrainian (KOI8-U); Cyrillic (KOI8-U).

iso_8859_1

ISO 8859-1 Latin 1; Western European (ISO).

iso_8859_2

ISO 8859-2 Central European; Central European (ISO).

iso_8859_3

ISO 8859-3 Latin 3.

iso_8859_4

ISO 8859-4 Baltic.

iso_8859_5

ISO 8859-5 Cyrillic.

iso_8859_6

ISO 8859-6 Arabic.

iso_8859_7

ISO 8859-7 Greek.

iso_8859_8

ISO 8859-8 Hebrew; Hebrew (ISO-Visual).

iso_8859_9

ISO 8859-9 Turkish.

iso_8859_13

ISO 8859-13 Estonian.

iso_8859_15

ISO 8859-15 Latin 9.

x_Europa

Europa 3.

iso_8859_8_i

ISO 8859-8 Hebrew; Hebrew (ISO-Logical).

iso_2022_jp_1

ISO 2022 Japanese with no halfwidth Katakana; Japanese (JIS).

csISO2022JP

ISO 2022 Japanese with halfwidth Katakana; Japanese (JIS-Allow 1 byte Kana).

iso_2022_jp_2

ISO 2022 Japanese JIS X 0201-1989; Japanese (JIS-Allow 1 byte Kana - SO/SI).

iso_2022_kr

ISO 2022 Korean.

x_cp50227

ISO 2022 Simplified Chinese; Chinese Simplified (ISO 2022).

iso_2022_ch

ISO 2022 Traditional Chinese.

x_EBCDIC_JapaneseExtended

EBCDIC Japanese (Katakana) Extended.

x_EBCDIC_USCanadaJapanese

EBCDIC US-Canada and Japanese.

x_EBCDIC_KoreanExtendedAndKorean

EBCDIC Korean Extended and Korean.

x_EBCDIC_ChineseExtendedAndChinese

EBCDIC Simplified Chinese Extended and Simplified Chinese.

x_EBCDIC_ChineseExtended

EBCDIC Simplified Chinese.

x_EBCDIC_USCanadaChinese

EBCDIC US-Canada and Traditional Chinese.

x_EBCDIC_JapaneseExtendedAndJapanese

EBCDIC Japanese (Latin) Extended and Japanese.

euc_jp

EUC Japanese.

EUC_CN

EUC Simplified Chinese; Chinese Simplified (EUC).

euc_kr

EUC Korean.

EUC_TCN

EUC Traditional Chinese.

hz_gb_2312

HZ-GB2312 Simplified Chinese; Chinese Simplified (HZ).

GB18030

Windows XP and later: GB18030 Simplified Chinese (4 byte); Chinese Simplified (GB18030).

x_iscii_de

ISCII Devanagari.

x_iscii_be

ISCII Bangla.

x_iscii_ta

ISCII Tamil.

x_iscii_te

ISCII Telugu.

x_iscii_as

ISCII Assamese.

x_iscii_or

ISCII Odia.

x_iscii_ka

ISCII Kannada.

x_iscii_ma

ISCII Malayalam.

x_iscii_gu

ISCII Gujarati.

x_iscii_pa

ISCII Punjabi.

utf_7

Unicode (UTF-7).

utf_8

Unicode (UTF-8).

Invalid

Invalid code page.

Methods

impl CodePage

pub fn get_value(&self) -> u16

Returns the integral value of the CodePage .

impl CodePage

pub fn get_info(self) -> WinResult<CodePageInfo>

Returns a CodePageInfo object which contains information about the CodePage.

Examples

let info = CodePage::utf_8.get_info().unwrap();
println!("{}", info.name);

Errors

FromUtf8Error: Returned if an error occurs while converting to a character.

FromUtf16Error: Returned if an error occurs while converting to a character.

IoError: Returned if an OS error occurs.