Struct winconsole::console::OutputSettings

pub struct OutputSettings {
    pub ProcessedOutput: bool,
    pub WrapAtEol: bool,
    pub VirtualTerminalProcessing: bool,
    pub DisableNewlineAutoReturn: bool,
    pub LVBGridWorldwide: bool,
}

Settings related to console output.

See

GetConsoleMode.

Fields
ProcessedOutput: bool

Should output be processed for escape sequences?

WrapAtEol: bool

Should the console cursor wrap at the end of the line?

VirtualTerminalProcessing: bool

Should virtual terminal sequences be processed?

DisableNewlineAutoReturn: bool

Should the console cursor return to the the first column on newline?

LVBGridWorldwide: bool

Should LVB attribute flags be enabled?

Methods

impl OutputSettings

pub fn new() -> OutputSettings

Creates a new OutputSettings object with all fields set to false.