Module winconsole::window

Contains window-related functions, structs, and enums.

Structs

FlashInfo
Information used to flash the window.

Enums

DisplayState
The display state of the console window.

Functions

activate
Brings the window to the foreground and activates it, optionally displaying it.

flash
Flashes the console window caption and/or tray icon.

get_cursor_position
Returns the screen position of the cursor.

get_display_state
Returns the display state of the console window.

get_position
Returns the position of the console window in pixels.

get_size
Returns the size of the console window in pixels.

hide
Hides the console window.

is_active
Is the window the active window?

is_maximized
Is the window maximized?

is_minimized
Is the window minimized?

is_visible
Is the window visible?

maximize
Maximizes and activates the console window.

minimize
Minimizes the console window.

restore
Restores and activates the console window.

set_cursor_position
Sets the screen position of the cursor.

set_display_state
Sets the display state of the console window.

set_position
Sets the position of the console window.

set_size
Sets the size of the console window.

show
Shows the console window.

stop_flash
Stops flashing the console window caption and/or tray icon.